+++ gazelle/python/testdata/binary_targets/bar_binary.py
@@ -1,0 +1,4 @@
+import bar
+
+if __name__ == "__main__":
+    pass