+++ gazelle/python/testdata/binary_targets/name_main_test.py
@@ -1,0 +1,3 @@
+# This should make a py_test target because of the filename
+if __name__ == "__main__":
+    pass