+++ gazelle/python/testdata/binary_targets/single_quote_main.py
@@ -1,0 +1,4 @@
+import bar_test
+
+if __name__ == '__main__':
+    pass